requirements.txt
asgiref==3.6.0
Django==4.2
django-etc==1.4.0
django-hitcount==1.3.5
Pillow==9.5.0
sqlparse==0.4.4
